Adding support for pubsub emulator in system test.
